+++ yadma/src/download/getdownloadinfo.h
@@ -1,8 +1,9 @@
 #ifndef GETDOWNLOAD_INFO_H
 #define GETDOWNLOAD_INFO_H
 
+#include <inttypes.h>
 #include <string>
 
-bool GetDownloadInfo( const std::string& Url, std::string& Cookie, std::string& Final, std::string& FileName, size_t& FileSize, bool Prefixed );
+bool GetDownloadInfo( const std::string& Url, std::string& Cookie, std::string& Final, std::string& FileName, int64_t& FileSize, bool Prefixed );
 
 #endif
